# adapter-service/openapi.team.v1.yaml
openapi: 3.0.3
info:
  title: ScoutX Adapter API
  version: 1.0.0
  description: |
    OpenAPI-like Referenz für Team- und Admin-Endpunkte des ScoutX Adapter-Service.
    Diese Spezifikation priorisiert produktionsrelevante Flows (Auth, Team-State, Planung, Notifications, Admin-Diagnostics).
servers:
  - url: http://127.0.0.1:8787
security:
  - bearerAuth: []
paths:
...
  /api/team/auth/login:
    post:
      summary: Team login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, password]
              properties:
                userId: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamStatePayload'
        '401':
          description: Invalid credentials
...
  /api/team/audit-log:
    get:
      summary: List audit log entries with optional actor/action filter
      parameters:
        - in: query
          name: actorId
          schema:
            type: string
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  count: { type: integer }
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        actorId: { type: string }
                        action: { type: string }
                        title: { type: string }
                        message: { type: string }
                        createdAt: { type: string, format: date-time }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    TeamStatePayload:
      type: object
      properties:
        ok: { type: boolean }
        user:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            role: { type: string }
            teamId: { type: string }
        team:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        manualGames:
          type: array
          items:
            $ref: '#/components/schemas/Game'
        observations:
          type: array
          items:
            $ref: '#/components/schemas/Observation'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        feedItems:
          type: array
          items:
            $ref: '#/components/schemas/FeedItem'
...
    Game:
      type: object
      properties:
        id: { type: string }
        source: { type: string }
        date: { type: string }
        time: { type: string }
        home: { type: string }
        away: { type: string }
        venue: { type: string }
        status: { type: string }
        note: { type: string }
    Observation:
      type: object
      properties:
        id: { type: string }
        gameId: { type: string }
        scoutId: { type: string }
        status: { type: string }
        note: { type: string }
        reportId: { type: string }
        reportUrl: { type: string }
    Notification:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        body: { type: string }
        unread: { type: boolean }
        createdAt: { type: string, format: date-time }
    FeedItem:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        createdAt: { type: string, format: date-time }
        actorId: { type: string }
        title: { type: string }
        body: { type: string }